test(active-directory): expose test helpers
