+++ packages/ubuntu_desktop_installer/test/active_directory/test_active_directory.mocks.dart
@@ -1,0 +1,335 @@
+// Mocks generated by Mockito 5.4.0 from annotations
+// in ubuntu_desktop_installer/test/active_directory/test_active_directory.dart.
+// Do not manually edit this file.
+
+// ignore_for_file: no_leading_underscores_for_library_prefixes
+import 'dart:async' as _i4;
+import 'dart:ui' as _i6;
+
+import 'package:mockito/mockito.dart' as _i1;
+import 'package:subiquity_client/subiquity_client.dart' as _i2;
+import 'package:ubuntu_desktop_installer/pages/active_directory/active_directory_model.dart'
+    as _i5;
+import 'package:ubuntu_desktop_installer/services.dart' as _i3;
+import 'package:ubuntu_wizard/src/utils/url_launcher.dart' as _i7;
+
+// ignore_for_file: type=lint
+// ignore_for_file: avoid_redundant_argument_values
+// ignore_for_file: avoid_setters_without_getters
+// ignore_for_file: comment_references
+// ignore_for_file: implementation_imports
+// ignore_for_file: invalid_use_of_visible_for_testing_member
+// ignore_for_file: prefer_const_constructors
+// ignore_for_file: unnecessary_parenthesis
+// ignore_for_file: camel_case_types
+// ignore_for_file: subtype_of_sealed_class
+
+class _FakeAdConnectionInfo_0 extends _i1.SmartFake
+    implements _i2.AdConnectionInfo {
+  _FakeAdConnectionInfo_0(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
+/// A class which mocks [ActiveDirectoryService].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockActiveDirectoryService extends _i1.Mock
+    implements _i3.ActiveDirectoryService {
+  MockActiveDirectoryService() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i4.Future<bool> hasSupport() => (super.noSuchMethod(
+        Invocation.method(
+          #hasSupport,
+          [],
+        ),
+        returnValue: _i4.Future<bool>.value(false),
+      ) as _i4.Future<bool>);
+  @override
+  _i4.Future<_i2.AdConnectionInfo> getConnectionInfo() => (super.noSuchMethod(
+        Invocation.method(
+          #getConnectionInfo,
+          [],
+        ),
+        returnValue:
+            _i4.Future<_i2.AdConnectionInfo>.value(_FakeAdConnectionInfo_0(
+          this,
+          Invocation.method(
+            #getConnectionInfo,
+            [],
+          ),
+        )),
+      ) as _i4.Future<_i2.AdConnectionInfo>);
+  @override
+  _i4.Future<void> setConnectionInfo(_i2.AdConnectionInfo? info) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #setConnectionInfo,
+          [info],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<List<_i2.AdDomainNameValidation>> checkDomainName(
+          String? domain) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #checkDomainName,
+          [domain],
+        ),
+        returnValue: _i4.Future<List<_i2.AdDomainNameValidation>>.value(
+            <_i2.AdDomainNameValidation>[]),
+      ) as _i4.Future<List<_i2.AdDomainNameValidation>>);
+  @override
+  _i4.Future<_i2.AdAdminNameValidation> checkAdminName(String? admin) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #checkAdminName,
+          [admin],
+        ),
+        returnValue: _i4.Future<_i2.AdAdminNameValidation>.value(
+            _i2.AdAdminNameValidation.OK),
+      ) as _i4.Future<_i2.AdAdminNameValidation>);
+  @override
+  _i4.Future<_i2.AdPasswordValidation> checkPassword(String? password) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #checkPassword,
+          [password],
+        ),
+        returnValue: _i4.Future<_i2.AdPasswordValidation>.value(
+            _i2.AdPasswordValidation.OK),
+      ) as _i4.Future<_i2.AdPasswordValidation>);
+  @override
+  _i4.Future<_i2.AdDomainNameValidation> pingDomainController(String? domain) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #pingDomainController,
+          [domain],
+        ),
+        returnValue: _i4.Future<_i2.AdDomainNameValidation>.value(
+            _i2.AdDomainNameValidation.OK),
+      ) as _i4.Future<_i2.AdDomainNameValidation>);
+  @override
+  _i4.Future<_i2.AdJoinResult> getJoinResult({bool? wait = true}) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getJoinResult,
+          [],
+          {#wait: wait},
+        ),
+        returnValue: _i4.Future<_i2.AdJoinResult>.value(_i2.AdJoinResult.OK),
+      ) as _i4.Future<_i2.AdJoinResult>);
+  @override
+  _i4.Future<void> markConfigured() => (super.noSuchMethod(
+        Invocation.method(
+          #markConfigured,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+}
+
+/// A class which mocks [ActiveDirectoryModel].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockActiveDirectoryModel extends _i1.Mock
+    implements _i5.ActiveDirectoryModel {
+  MockActiveDirectoryModel() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  String get domainName => (super.noSuchMethod(
+        Invocation.getter(#domainName),
+        returnValue: '',
+      ) as String);
+  @override
+  bool get isDomainNameValid => (super.noSuchMethod(
+        Invocation.getter(#isDomainNameValid),
+        returnValue: false,
+      ) as bool);
+  @override
+  String get adminName => (super.noSuchMethod(
+        Invocation.getter(#adminName),
+        returnValue: '',
+      ) as String);
+  @override
+  bool get isAdminNameValid => (super.noSuchMethod(
+        Invocation.getter(#isAdminNameValid),
+        returnValue: false,
+      ) as bool);
+  @override
+  String get password => (super.noSuchMethod(
+        Invocation.getter(#password),
+        returnValue: '',
+      ) as String);
+  @override
+  bool get isPasswordValid => (super.noSuchMethod(
+        Invocation.getter(#isPasswordValid),
+        returnValue: false,
+      ) as bool);
+  @override
+  bool get isValid => (super.noSuchMethod(
+        Invocation.getter(#isValid),
+        returnValue: false,
+      ) as bool);
+  @override
+  bool get hasListeners => (super.noSuchMethod(
+        Invocation.getter(#hasListeners),
+        returnValue: false,
+      ) as bool);
+  @override
+  bool get isDisposed => (super.noSuchMethod(
+        Invocation.getter(#isDisposed),
+        returnValue: false,
+      ) as bool);
+  @override
+  _i4.Future<void> setDomainName(String? value) => (super.noSuchMethod(
+        Invocation.method(
+          #setDomainName,
+          [value],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<void> validateDomainName() => (super.noSuchMethod(
+        Invocation.method(
+          #validateDomainName,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<void> pingDomainController() => (super.noSuchMethod(
+        Invocation.method(
+          #pingDomainController,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<void> setAdminName(String? value) => (super.noSuchMethod(
+        Invocation.method(
+          #setAdminName,
+          [value],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<void> validateAdminName() => (super.noSuchMethod(
+        Invocation.method(
+          #validateAdminName,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<void> setPassword(String? value) => (super.noSuchMethod(
+        Invocation.method(
+          #setPassword,
+          [value],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<void> validatePassword() => (super.noSuchMethod(
+        Invocation.method(
+          #validatePassword,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<void> init() => (super.noSuchMethod(
+        Invocation.method(
+          #init,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<void> save() => (super.noSuchMethod(
+        Invocation.method(
+          #save,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<_i2.AdJoinResult> getJoinResult() => (super.noSuchMethod(
+        Invocation.method(
+          #getJoinResult,
+          [],
+        ),
+        returnValue: _i4.Future<_i2.AdJoinResult>.value(_i2.AdJoinResult.OK),
+      ) as _i4.Future<_i2.AdJoinResult>);
+  @override
+  void addListener(_i6.VoidCallback? listener) => super.noSuchMethod(
+        Invocation.method(
+          #addListener,
+          [listener],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void removeListener(_i6.VoidCallback? listener) => super.noSuchMethod(
+        Invocation.method(
+          #removeListener,
+          [listener],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void dispose() => super.noSuchMethod(
+        Invocation.method(
+          #dispose,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void notifyListeners() => super.noSuchMethod(
+        Invocation.method(
+          #notifyListeners,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
+}
+
+/// A class which mocks [UrlLauncher].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockUrlLauncher extends _i1.Mock implements _i7.UrlLauncher {
+  MockUrlLauncher() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i4.Future<bool> launchUrl(String? url) => (super.noSuchMethod(
+        Invocation.method(
+          #launchUrl,
+          [url],
+        ),
+        returnValue: _i4.Future<bool>.value(false),
+      ) as _i4.Future<bool>);
+}
